+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Change Log
 
 ## [Unreleased]
+### Removed
+- Remove experimental useReduxStateMapped
+### Changed
+- useLayoutEffect and keep latest state after update (see #20)
 ### Added
 - New useReduxSelectors
-### Changed
-- Removed useReduxStateMapped
 
 ## [1.8.0] - 2019-04-02
 ### Changed